+++ docs/shared/route-manifest.gen.ts
@@ -1,11 +1,11 @@
 /* eslint-disable */
 import { createRouteNode } from '@richie-router/core';
-import { headTagSchema } from './head-tag-schema.ts';
+import { routerSchema } from './router-schema.ts';
 
-const __rootRoute = createRouteNode('__root__', { head: 'docs-shell' }, { isRoot: true })._setSearchSchema((headTagSchema as any)['docs-shell']?.searchSchema as never);
-const IndexRoute = createRouteNode('/', {}).update({ id: '/', path: '/', getParentRoute: () => __rootRoute } as const);
-const DocsSplatslugRoute = createRouteNode('/docs/$slug', { head: 'document-page' }).update({ id: '/docs/$slug', path: '/docs/$slug', getParentRoute: () => __rootRoute } as const)._setSearchSchema((headTagSchema as any)['document-page']?.searchSchema as never);
-const SearchRoute = createRouteNode('/search', { head: 'docs-search' }).update({ id: '/search', path: '/search', getParentRoute: () => __rootRoute } as const)._setSearchSchema((headTagSchema as any)['docs-search']?.searchSchema as never);
+const __rootRoute = createRouteNode('__root__', {}, { isRoot: true })._setSearchSchema((routerSchema as any)['__root__']?.searchSchema as never)._setServerHead((routerSchema as any)['__root__']?.serverHead);
+const IndexRoute = createRouteNode('/', {}).update({ id: '/', path: '/', getParentRoute: () => __rootRoute } as const)._setSearchSchema((routerSchema as any)['/']?.searchSchema as never)._setServerHead((routerSchema as any)['/']?.serverHead);
+const DocsSplatslugRoute = createRouteNode('/docs/$slug', {}).update({ id: '/docs/$slug', path: '/docs/$slug', getParentRoute: () => __rootRoute } as const)._setSearchSchema((routerSchema as any)['/docs/$slug']?.searchSchema as never)._setServerHead((routerSchema as any)['/docs/$slug']?.serverHead);
+const SearchRoute = createRouteNode('/search', {}).update({ id: '/search', path: '/search', getParentRoute: () => __rootRoute } as const)._setSearchSchema((routerSchema as any)['/search']?.searchSchema as never)._setServerHead((routerSchema as any)['/search']?.serverHead);
 
 
 const __rootRouteChildren = {