+++ scripts/generate-demo-routes.ts
@@ -3,7 +3,7 @@ import { generateRouteTree } from '@richie-router/tooling';
 
 await generateRouteTree({
   routesDir: path.resolve('demo/frontend/routes'),
-  headTagSchema: path.resolve('demo/shared/head-tag-schema.ts'),
-  output: path.resolve('demo/shared/route-tree.gen.ts'),
+  routerSchema: path.resolve('demo/shared/router-schema.ts'),
+  output: path.resolve('demo/frontend/route-tree.gen.ts'),
   manifestOutput: path.resolve('demo/shared/route-manifest.gen.ts'),
 });