+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @richie-router/core
 
+## 0.1.2
+
+### Patch Changes
+
+- change structure a bit to get greater frontend/backend separation
+
 ## 0.1.1
 
 ### Patch Changes