+++ packages/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @richie-router/react
 
+## 0.1.2
+
+### Patch Changes
+
+- change structure a bit to get greater frontend/backend separation
+- Updated dependencies
+  - @richie-router/core@0.1.2
+
 ## 0.1.1
 
 ### Patch Changes