+++ packages/server/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @richie-router/server
 
+## 0.1.3
+
+### Patch Changes
+
+- change structure a bit to get greater frontend/backend separation
+- Updated dependencies
+  - @richie-router/core@0.1.2
+
 ## 0.1.2
 
 ### Patch Changes